Fix tags and methods
An HTML tag ends at the closing delimiter of its string, even if the
tag is not properly ended. For example,

    '<font color='.length()

evaluates to 12. It is not an error.

HTML attributes need not be terminated within a tag. For example,

    "<font color='purple>igram</font>";

does not display anything, because the attribute has the value
"purple>igram</font>" and there is no text to display.

A property name immediately following a method definition was lexed as
Name, instead of Name.Variable.