Speed up the regex for whitespace and comments
Premature optimization strikes again!

This regex was unbearably slow in some situations. The relevant part
was `r'(\s+)*'`, where the two quantifiers made it keep backtracking.
This was seldom a problem: although it is legal to have many spaces
and/or comments between, say, a label and a colon, it is rare. It was
only of practical concern where an identifier which was not a label,
but which appeared in a position where a label was legal, appeared
before many spaces. For example, from `en_us.t`:

                            else
                                cur = cur.substr(2);

It took forever to figure out that `else` is not a label, because no
colon follows it, no matter how much it backtracks.

Removing the `+` quantifier resolves the issue.